Checkpoint seems to work now
After some tweaks it should be working now.
On download resume, page counter is not stored so it will show -> 'page = 1' but thats the number of page for current download. It should continue instead of checking all previously downloaded contents.
On retry 1/5 it kinda breaks but you push Ctrl+C to stop and.. yeah thats about it.